+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 # Unreleased
 
-- Fixed a PHP error that could occur when saving a discount. ([#3538](https://github.com/craftcms/commerce/issues/3538)) 
+- Fixed a PHP error that could occur when saving a discount. ([#3538](https://github.com/craftcms/commerce/issues/3538))
+- Added `craft\commerce\behaviors\CustomerAddressBehavior::defineFields()`.
+- Added `craft\commerce\behaviors\CustomerBehavior::defineFields()`.
 
 ## 5.0.9 - 2024-06-05